refactor: moves CborSerializer
